Attributes g:d:vomsGroupName and g:d:vomsRoles moved to group-resource.
This change was done because we could not map one group to different groups on different
VOMS servers or to same VOMS server but with different VO.
